+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @berkayorhan/stackkit
 
+## 0.2.0
+
+### Minor Changes
+
+- Add typed native initializer contracts for official CLI delegation.
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @berkayorhan/stackkit-core@0.2.0
+  - @berkayorhan/stackkit-registry@0.2.0
+  - @berkayorhan/stackkit-schemas@0.2.0
+
 ## 0.1.2
 
 ### Patch Changes