+++ packages/schemas/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @berkayorhan/stackkit-schemas
 
+## 0.2.0
+
+### Minor Changes
+
+- Add typed native initializer contracts for official CLI delegation.
+
 ## 0.1.2
 
 ### Patch Changes